+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Updated the domain of the question. Reduced the number of attributes from 12 to 8, integrated _sensitive question_ option to question domain.
 - Updated the question asking flow/process on telegram according to M46 specification.
 
+:house: Internal
+- Updated WeNet Common to version 5.0.1
+
 ### 3.0.0
 
 :rocket: New features